Make common modal component for system picking and executing some function with selected ids
